requirements.txt
keyboard>=0.13.5
pyautogui>=0.9.54
pystray>=0.19.5
Pillow>=10.0.0
